requirements.txt
python-dotenv~=0.20.0
aiogram~=2.21
requests~=2.28.0
beautifulsoup4~=4.11.1
Columnar~=1.4.1
SQLAlchemy~=1.4.39
loguru~=0.6.0
PlexAPI~=4.12.1
python-qbittorrent~=0.4.2
aioschedule~=0.5.2
APScheduler~=3.9.1.post1